Fix the "Setting properties ... considered harmful" console warning
This warning is displayed by LitElement when property change handler
(an implementation of `_propertiesChanged` method) triggers another
property change (for a "managed" property registered in the `properties`
hash).

To solve this, the intermediate `_expandedItem` property, which was
changed in response to `expandedItem` property change, was removed.
The `_handleExpandedItemChange` method, which relied on setting this
property to trigger the element UI update in the middle of method's '
lifecycle was reworked to trigger the UI update by invoking the
(injected) `super._propertiesChanged` method in the respective place.